+++ CHANGELOG.md
@@ -13,6 +13,8 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 ### Fixed
 - Opening an already open Bolt database should not cause sensu-agent to hang
 indefinitely.
+- Added the `APIKey` resource and HTTP API support for POST, GET, PATCH and DELETE.
+- Added sensuctl commands to manage the `APIKey` resource.
 
 ## [5.14.0] - 2019-10-08